(GH-2320) Update template project to reflect module workflow
This updates the template project and described directories to include
the new module workflow. This includes adding `modules` configuration to
`bolt-project.yaml` that pulls in the facts module from the forge and
the service module from Github. I chose these modules as they are
already shipped with Bolt, so users who copy-paste won't be pulling in
unecessary extra modules, and these versions should be compatible with
the versions used in Bolt for the most part.

This also updates the readme to reference `modules/` as the place to
develop custom modules, describes the Puppetfile as a lockfile, and
includes `.modules/` as the location for installed modules.

Closes puppetlabs/bolt#2320